Allow custom levels to use 2 billion tile numbers once again
Ever since tilesheets got expanded, custom levels could use as many
tiles as they wanted, as long as it fit under the 32-bit signed integer
limit.

Until 6c85fae339f83f032230eafaf8ee7742f03dbbac happened and they were
reduced to 32,767 tiles.

So I'm being generous again and changing the type of the contents array
(in mapclass and editorclass) back to int. This won't affect the
existing tilemaps of the main game, they'll still stay short arrays. But
it means level makers can use 2 billion tiles once again.

+++ desktop_version/src/Map.h
@@ -91,7 +91,7 @@ public:
     int roomdeaths[20 * 20];
     int roomdeathsfinal[20 * 20];
     static const int areamap[20 * 20];
-    short contents[40 * 30];
+    int contents[40 * 30];
     bool explored[20 * 20];
     int vmult[30];
