+++ desktop_version/src/editor.cpp
@@ -539,7 +539,7 @@ void editorclass::getlin(const enum textmode mode, const std::string& prompt, st
     oldenttext = key.keybuffer;
 }
 
-const short* editorclass::loadlevel( int rxi, int ryi )
+const int* editorclass::loadlevel( int rxi, int ryi )
 {
     //Set up our buffer array to be picked up by mapclass
     rxi -= 100;
@@ -549,7 +549,7 @@ const short* editorclass::loadlevel( int rxi, int ryi )
     if(rxi>=mapwidth)rxi-=mapwidth;
     if(ryi>=mapheight)ryi-=mapheight;
 
-    static short result[1200];
+    static int result[1200];
 
     for (int j = 0; j < 30; j++)
     {
