+++ desktop_version/src/editor.h
@@ -132,7 +132,7 @@ class editorclass{
 
   void reset(void);
   void getlin(const enum textmode mode, const std::string& prompt, std::string* ptr);
-  const short* loadlevel(int rxi, int ryi);
+  const int* loadlevel(int rxi, int ryi);
 
   int gettileidx(
     const int rx,
@@ -217,7 +217,7 @@ class editorclass{
   std::vector<std::string> getLevelDirFileNames( );
   static const int maxwidth = 20, maxheight = 20; //Special; the physical max the engine allows
   static const int numrooms = maxwidth * maxheight;
-  short contents[40 * 30 * numrooms];
+  int contents[40 * 30 * numrooms];
   int vmult[30 * maxheight];
   int numtrinkets(void);
   int numcrewmates(void);
